나누셈 고쳐야함 ㅠㅠ

+++ web.py
@@ -15,7 +15,7 @@ def calcExpr():
       expr = request.form['Expression']
       expr_range = request.form['Range']
       if expr_range == '': expr_range = '-1,1'
-      expr_range = list(map(int,expr_range.split(',')))
+      expr_range = list(map(float,expr_range.split(',')))
       pics, canonicalization, partial_derivatives, domain = test(expr,expr_range)
       return render_template('image.html',
                              pics = pics,
